Fish Spawner That Will Not Spawn Fish
I made a spawner and three game objects representing the three different sizes. I made the game objects prefabs so I could reference them in the script. I made the script for the spawner, but I don't know if it will work because I couldn't figure out how to test it.